app/requirements.txt
Django
gunicorn
django-cors-headers
django-s3-storage
djangorestframework
# aws-psycopg2
psycopg2-binary
progressbar2
boto3
djangorestframework-jwt
PyJWT==1.7.1
pytz
zappa-django-utils
pandas
numpy
sqlalchemy
shortuuid
psutil